refactor : common package name update
