+++ src/main/java/kr/co/springbootweeklymission/global/error/exception/WrongVoucherPolicyException.java
@@ -1,6 +1,6 @@
-package kr.co.springbootweeklymission.common.error.exception;
+package kr.co.springbootweeklymission.global.error.exception;
 
-import kr.co.springbootweeklymission.common.response.ResponseStatus;
+import kr.co.springbootweeklymission.global.response.ResponseStatus;
 
 public class WrongVoucherPolicyException extends RuntimeException {
     public WrongVoucherPolicyException(ResponseStatus responseStatus) {
