+++ src/main/java/kr/co/springbootweeklymission/voucher/application/VoucherService.java
@@ -1,7 +1,7 @@
 package kr.co.springbootweeklymission.voucher.application;
 
-import kr.co.springbootweeklymission.common.error.exception.NotFoundException;
-import kr.co.springbootweeklymission.common.response.ResponseStatus;
+import kr.co.springbootweeklymission.global.error.exception.NotFoundException;
+import kr.co.springbootweeklymission.global.response.ResponseStatus;
 import kr.co.springbootweeklymission.voucher.domain.entity.Voucher;
 import kr.co.springbootweeklymission.voucher.domain.repository.VoucherRepository;
 import kr.co.springbootweeklymission.voucher.presentation.dto.request.VoucherReqDTO;
