+++ src/main/java/kr/co/springbootweeklymission/voucher/domain/repository/InFileVoucherRepository.java
@@ -1,9 +1,9 @@
 package kr.co.springbootweeklymission.voucher.domain.repository;
 
-import kr.co.springbootweeklymission.common.error.exception.FileIOException;
-import kr.co.springbootweeklymission.common.error.exception.NotSupportedException;
-import kr.co.springbootweeklymission.common.response.ResponseStatus;
-import kr.co.springbootweeklymission.common.util.FileConverter;
+import kr.co.springbootweeklymission.global.error.exception.FileIOException;
+import kr.co.springbootweeklymission.global.error.exception.NotSupportedException;
+import kr.co.springbootweeklymission.global.response.ResponseStatus;
+import kr.co.springbootweeklymission.global.util.FileConverter;
 import kr.co.springbootweeklymission.voucher.domain.entity.Voucher;
 import kr.co.springbootweeklymission.voucher.presentation.dto.response.VoucherResDTO;
 import org.springframework.context.annotation.Profile;
