+++ src/test/java/kr/co/springbootweeklymission/voucher/domain/entity/VoucherTest.java
@@ -1,7 +1,7 @@
 package kr.co.springbootweeklymission.voucher.domain.entity;
 
-import kr.co.springbootweeklymission.common.error.exception.WrongVoucherPolicyException;
-import kr.co.springbootweeklymission.common.response.ResponseStatus;
+import kr.co.springbootweeklymission.global.error.exception.WrongVoucherPolicyException;
+import kr.co.springbootweeklymission.global.response.ResponseStatus;
 import kr.co.springbootweeklymission.voucher.domain.model.VoucherPolicy;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
